requirements.txt
# === DEPENDENCIAS PRINCIPALES ===
mariadb==1.1.8
python-dotenv==1.0.0
tabulate==0.9.0
colorama==0.4.6
pyfiglet==0.8.post1

# === MANEJO DE FECHAS/HORAS ===
python-dateutil==2.8.2